pose_tokenizer/requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Data processing
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
pillow>=8.3.0

# Machine learning utilities
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
pandas>=1.3.0

# Configuration and logging
pyyaml>=6.0
omegaconf>=2.1.0
tensorboard>=2.8.0
wandb>=0.12.0
tqdm>=4.62.0

# Model export and optimization
onnx>=1.12.0
onnxruntime>=1.12.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Distributed training
accelerate>=0.16.0

# Optional dependencies for advanced features
# Uncomment if needed:
# transformers>=4.20.0
# timm>=0.6.0
# einops>=0.4.0